Rearrange components
